+++ orb/analytics/urls.py
@@ -11,5 +11,5 @@ urlpatterns = patterns('',
     url(r'^map/$', 'orb.analytics.views.map_view', name="orb_analytics_map"),
     url(r'^tag/(?P<id>\d+)/$', 'orb.analytics.views.tag_view', name="orb_analytics_tag"),
     url(r'^tag/(?P<id>\d+)/download/(?P<year>\d+)/(?P<month>\d+)/$', 'orb.analytics.views.tag_download', name="orb_analytics_download"),
-
+    url(r'^resource/(?P<id>\d+)/$', 'orb.analytics.views.resource_view', name="orb_analytics_resource"),
 )